@@ -1,26 +1,26 @@
 {
-  "_from": "compressible@~2.0.14",
-  "_id": "compressible@2.0.16",
+  "_from": "compressible@~2.0.16",
+  "_id": "compressible@2.0.17",
   "_inBundle": false,
-  "_integrity": "sha512-JQfEOdnI7dASwCuSPWIeVYwc/zMsu/+tRhoUvEfXz2gxOA2DNjmG5vhtFdBlhWPPGo+RdT9S3tgc/uH5qgDiiA==",
+  "_integrity": "sha512-BGHeLCK1GV7j1bSmQQAi26X+GgWcTjLr/0tzSvMCl3LH1w1IJ4PFSPoV5316b30cneTziC+B1a+3OjoSUcQYmw==",
   "_location": "/compressible",
   "_phantomChildren": {},
   "_requested": {
     "type": "range",
     "registry": true,
-    "raw": "compressible@~2.0.14",
+    "raw": "compressible@~2.0.16",
     "name": "compressible",
     "escapedName": "compressible",
-    "rawSpec": "~2.0.14",
+    "rawSpec": "~2.0.16",
     "saveSpec": null,
-    "fetchSpec": "~2.0.14"
+    "fetchSpec": "~2.0.16"
   },
   "_requiredBy": [
     "/compression"
   ],
-  "_resolved": "https://registry.npmjs.org/compressible/-/compressible-2.0.16.tgz",
-  "_shasum": "a49bf9858f3821b64ce1be0296afc7380466a77f",
-  "_spec": "compressible@~2.0.14",
+  "_resolved": "https://registry.npmjs.org/compressible/-/compressible-2.0.17.tgz",
+  "_shasum": "6e8c108a16ad58384a977f3a482ca20bff2f38c1",
+  "_spec": "compressible@~2.0.16",
   "_where": "/Users/air/Desktop/diploma2/client/node_modules/compression",
   "bugs": {
     "url": "https://github.com/jshttp/compressible/issues"
@@ -43,20 +43,20 @@
     }
   ],
   "dependencies": {
-    "mime-db": ">= 1.38.0 < 2"
+    "mime-db": ">= 1.40.0 < 2"
   },
   "deprecated": false,
   "description": "Compressible Content-Type / mime checking",
   "devDependencies": {
-    "eslint": "5.14.1",
+    "eslint": "5.15.2",
     "eslint-config-standard": "12.0.0",
     "eslint-plugin-import": "2.16.0",
     "eslint-plugin-markdown": "1.0.0",
     "eslint-plugin-node": "8.0.1",
     "eslint-plugin-promise": "4.0.1",
     "eslint-plugin-standard": "4.0.0",
-    "mocha": "5.2.0",
-    "nyc": "13.3.0"
+    "mocha": "6.1.4",
+    "nyc": "14.0.0"
   },
   "engines": {
     "node": ">= 0.6"
@@ -87,5 +87,5 @@
     "test-travis": "nyc --reporter=text npm test",
     "version": "node scripts/version-history.js && git add HISTORY.md"
   },
-  "version": "2.0.16"
+  "version": "2.0.17"
 }
