@@ -1,8 +1,8 @@
 {
   "_from": "compression@^1.5.2",
-  "_id": "compression@1.7.3",
+  "_id": "compression@1.7.4",
   "_inBundle": false,
-  "_integrity": "sha512-HSjyBG5N1Nnz7tF2+O7A9XUhyjru71/fwgNb7oIsEVHR0WShfs2tIS/EySLgiTe98aOK18YDlMXpzjCXY/n9mg==",
+  "_integrity": "sha512-jaSIDzP9pZVS4ZfQ+TzvtiWhdpFhE2RDHz8QJkpX9SIpLq88VueF5jJw6t+6CUQcAoA6t+x89MLrWAqpfDE8iQ==",
   "_location": "/compression",
   "_phantomChildren": {},
   "_requested": {
@@ -18,8 +18,8 @@
   "_requiredBy": [
     "/webpack-dev-server"
   ],
-  "_resolved": "https://registry.npmjs.org/compression/-/compression-1.7.3.tgz",
-  "_shasum": "27e0e176aaf260f7f2c2813c3e440adb9f1993db",
+  "_resolved": "https://registry.npmjs.org/compression/-/compression-1.7.4.tgz",
+  "_shasum": "95523eff170ca57c29a0ca41e6fe131f41e5bb8f",
   "_spec": "compression@^1.5.2",
   "_where": "/Users/air/Desktop/diploma2/client/node_modules/webpack-dev-server",
   "bugs": {
@@ -40,9 +40,9 @@
   "dependencies": {
     "accepts": "~1.3.5",
     "bytes": "3.0.0",
-    "compressible": "~2.0.14",
+    "compressible": "~2.0.16",
     "debug": "2.6.9",
-    "on-headers": "~1.0.1",
+    "on-headers": "~1.0.2",
     "safe-buffer": "5.1.2",
     "vary": "~1.1.2"
   },
@@ -50,16 +50,16 @@
   "description": "Node.js compression middleware",
   "devDependencies": {
     "after": "0.8.2",
-    "eslint": "4.19.1",
-    "eslint-config-standard": "11.0.0",
-    "eslint-plugin-import": "2.13.0",
-    "eslint-plugin-markdown": "1.0.0-beta.6",
-    "eslint-plugin-node": "6.0.1",
-    "eslint-plugin-promise": "3.8.0",
-    "eslint-plugin-standard": "3.1.0",
+    "eslint": "5.15.1",
+    "eslint-config-standard": "12.0.0",
+    "eslint-plugin-import": "2.16.0",
+    "eslint-plugin-markdown": "1.0.0",
+    "eslint-plugin-node": "7.0.1",
+    "eslint-plugin-promise": "4.0.1",
+    "eslint-plugin-standard": "4.0.0",
     "istanbul": "0.4.5",
-    "mocha": "2.5.3",
-    "supertest": "1.1.0"
+    "mocha": "6.0.2",
+    "supertest": "4.0.0"
   },
   "engines": {
     "node": ">= 0.8.0"
@@ -82,5 +82,5 @@
     "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --check-leaks --reporter dot",
     "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --check-leaks --reporter spec"
   },
-  "version": "1.7.3"
+  "version": "1.7.4"
 }
