@@ -1,8 +1,8 @@
 {
   "_from": "proxy-addr@~2.0.4",
-  "_id": "proxy-addr@2.0.4",
+  "_id": "proxy-addr@2.0.5",
   "_inBundle": false,
-  "_integrity": "sha512-5erio2h9jp5CHGwcybmxmVqHmnCBZeewlfJ0pex+UW7Qny7OOZXTtH56TGNyBizkgiOwhJtMKrVzDTeKcySZwA==",
+  "_integrity": "sha512-t/7RxHXPH6cJtP0pRG6smSr9QJidhB+3kXu0KgXnbGYMgzEnUxRQ4/LDdfOwZEMyIh3/xHb8PX3t+lfL9z+YVQ==",
   "_location": "/proxy-addr",
   "_phantomChildren": {},
   "_requested": {
@@ -18,8 +18,8 @@
   "_requiredBy": [
     "/express"
   ],
-  "_resolved": "https://registry.npmjs.org/proxy-addr/-/proxy-addr-2.0.4.tgz",
-  "_shasum": "ecfc733bf22ff8c6f407fa275327b9ab67e48b93",
+  "_resolved": "https://registry.npmjs.org/proxy-addr/-/proxy-addr-2.0.5.tgz",
+  "_shasum": "34cbd64a2d81f4b1fd21e76f9f06c8a45299ee34",
   "_spec": "proxy-addr@~2.0.4",
   "_where": "/Users/air/Desktop/diploma2/client/node_modules/express",
   "author": {
@@ -32,22 +32,23 @@
   "bundleDependencies": false,
   "dependencies": {
     "forwarded": "~0.1.2",
-    "ipaddr.js": "1.8.0"
+    "ipaddr.js": "1.9.0"
   },
   "deprecated": false,
   "description": "Determine address of proxied request",
   "devDependencies": {
     "beautify-benchmark": "0.2.4",
     "benchmark": "2.1.4",
-    "eslint": "4.19.1",
-    "eslint-config-standard": "11.0.0",
-    "eslint-plugin-import": "2.13.0",
-    "eslint-plugin-markdown": "1.0.0-beta.6",
-    "eslint-plugin-node": "6.0.1",
-    "eslint-plugin-promise": "3.8.0",
-    "eslint-plugin-standard": "3.1.0",
-    "mocha": "3.5.3",
-    "nyc": "10.3.2"
+    "deep-equal": "1.0.1",
+    "eslint": "5.16.0",
+    "eslint-config-standard": "12.0.0",
+    "eslint-plugin-import": "2.17.1",
+    "eslint-plugin-markdown": "1.0.0",
+    "eslint-plugin-node": "8.0.1",
+    "eslint-plugin-promise": "4.1.1",
+    "eslint-plugin-standard": "4.0.0",
+    "mocha": "6.1.3",
+    "nyc": "13.3.0"
   },
   "engines": {
     "node": ">= 0.10"
@@ -77,5 +78,5 @@
     "test-cov": "nyc --reporter=text npm test",
     "test-travis": "nyc --reporter=html --reporter=text npm test"
   },
-  "version": "2.0.4"
+  "version": "2.0.5"
 }
