@@ -1,8 +1,8 @@
 {
   "_from": "mime@^2.3.1",
-  "_id": "mime@2.4.0",
+  "_id": "mime@2.4.2",
   "_inBundle": false,
-  "_integrity": "sha512-ikBcWwyqXQSHKtciCcctu9YfPbFYZ4+gbHEmE0Q8jzcTYQg5dHCr3g2wwAZjPoJfQVXZq6KXAjpXOTf5/cjT7w==",
+  "_integrity": "sha512-zJBfZDkwRu+j3Pdd2aHsR5GfH2jIWhmL1ZzBoc+X+3JEti2hbArWcyJ+1laC1D2/U/W1a/+Cegj0/OnEU2ybjg==",
   "_location": "/webpack-dev-middleware/mime",
   "_phantomChildren": {},
   "_requested": {
@@ -18,8 +18,8 @@
   "_requiredBy": [
     "/webpack-dev-middleware"
   ],
-  "_resolved": "https://registry.npmjs.org/mime/-/mime-2.4.0.tgz",
-  "_shasum": "e051fd881358585f3279df333fe694da0bcffdd6",
+  "_resolved": "https://registry.npmjs.org/mime/-/mime-2.4.2.tgz",
+  "_shasum": "ce5229a5e99ffc313abac806b482c10e7ba6ac78",
   "_spec": "mime@^2.3.1",
   "_where": "/Users/air/Desktop/diploma2/client/node_modules/webpack-dev-middleware",
   "author": {
@@ -39,14 +39,15 @@
   "deprecated": false,
   "description": "A comprehensive library for mime-type mapping",
   "devDependencies": {
-    "chalk": "1.1.3",
-    "eslint": "^5.9.0",
-    "mime-db": "^1.37.0",
-    "mime-score": "1.0.1",
-    "mime-types": "2.1.15",
-    "mocha": "5.2.0",
-    "runmd": "1.0.1",
-    "standard-version": "^4.4.0"
+    "benchmark": "^2.1.4",
+    "chalk": "2.4.2",
+    "eslint": "5.16.0",
+    "mime-db": "1.38.0",
+    "mime-score": "1.1.2",
+    "mime-types": "2.1.22",
+    "mocha": "6.0.2",
+    "runmd": "1.2.1",
+    "standard-version": "5.0.2"
   },
   "engines": {
     "node": ">=4.0.0"
@@ -63,10 +64,11 @@
     "type": "git"
   },
   "scripts": {
+    "benchmark": "node src/benchmark.js",
     "md": "runmd --watch --output README.md src/README_js.md",
     "prepare": "node src/build.js && runmd --output README.md src/README_js.md",
     "release": "standard-version",
     "test": "mocha src/test.js"
   },
-  "version": "2.4.0"
+  "version": "2.4.2"
 }
