@@ -5,88 +5,105 @@ import {AppComponent} from './app.component';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
 import {
   MAT_LABEL_GLOBAL_OPTIONS,
-  MatButtonModule,
-  MatCardModule,
-  MatCheckboxModule,
-  MatChipsModule,
-  MatDialogModule,
+  MatProgressSpinnerModule,
+  MatPaginatorModule,
   MatFormFieldModule,
   MatGridListModule,
-  MatIconModule,
+  MatCheckboxModule,
+  MatToolbarModule,
+  MatSidenavModule,
+  MatStepperModule,
+  MatDialogModule,
+  MatSelectModule,
+  MatButtonModule,
   MatInputModule,
+  MatChipsModule,
+  MatTableModule,
+  MatCardModule,
+  MatIconModule,
   MatListModule,
   MatMenuModule,
-  MatPaginatorModule,
-  MatProgressSpinnerModule,
-  MatSelectModule,
-  MatSidenavModule,
   MatSortModule,
-  MatStepperModule,
-  MatTableModule,
-  MatToolbarModule
 } from '@angular/material';
-import {NavComponent} from './components/nav/nav.component';
-import {LayoutModule} from '@angular/cdk/layout';
-import {JobsComponent} from './components/jobs/jobs.component';
-import {HTTP_INTERCEPTORS, HttpClient, HttpClientModule} from '@angular/common/http';
-import {FormsModule, ReactiveFormsModule} from '@angular/forms';
-import {CategoriesComponent} from './components/categories/categories.component';
+import {
+  HTTP_INTERCEPTORS,
+  HttpClientModule,
+  HttpClient,
+} from '@angular/common/http';
+
 import {CategoryEditComponent} from './components/categories/category-edit/category-edit.component';
-import {ToastrModule} from 'ngx-toastr';
+import {PageNotFoundComponent} from './components/errors/page-not-found/page-not-found.component';
+import {WorkersControlComponent} from './components/workers-control/workers-control.component';
+import {WorkerEditComponent} from './components/workers/worker-edit/worker-edit.component';
+import {CategoriesComponent} from './components/categories/categories.component';
 import {JobEditComponent} from './components/jobs/job-edit/job-edit.component';
 import {WorkersComponent} from './components/workers/workers.component';
-import {WorkerEditComponent} from './components/workers/worker-edit/worker-edit.component';
-import {TasksComponent} from './components/tasks/tasks.component';
+import {TranslateLoader, TranslateModule} from '@ngx-translate/core';
 import {ThreadComponent} from './components/thread/thread.component';
-import {WorkersControlComponent} from './components/workers-control/workers-control.component';
-import {PageNotFoundComponent} from './components/errors/page-not-found/page-not-found.component';
+import {TasksComponent} from './components/tasks/tasks.component';
+import {FormsModule, ReactiveFormsModule} from '@angular/forms';
+import {TranslateHttpLoader} from '@ngx-translate/http-loader';
+import {JobsComponent} from './components/jobs/jobs.component';
+import {NavComponent} from './components/nav/nav.component';
 import {ErrorInterceptor} from './interceptors/error';
 import {JwtInterceptor} from './interceptors/jwt';
-import { CommonModule } from '@angular/common';
+import {LayoutModule} from '@angular/cdk/layout';
+import {CommonModule} from '@angular/common';
+import {ToastrModule} from 'ngx-toastr';
+import {Observable} from 'rxjs';
+
+
+export class AccountTranslationLoader implements TranslateLoader {
+
+  constructor(private http: HttpClient) {}
+
+  getTranslation(lang: string): Observable<any> {
+    return this.http.get(`./assets/i18n/${lang}.json`);
+  }
+}
 
 @NgModule({
   declarations: [
-    AppComponent,
-    NavComponent,
-    JobsComponent,
-    CategoriesComponent,
+    WorkersControlComponent,
+    PageNotFoundComponent,
     CategoryEditComponent,
+    CategoriesComponent,
+    WorkerEditComponent,
     JobEditComponent,
     WorkersComponent,
-    WorkerEditComponent,
-    TasksComponent,
     ThreadComponent,
-    WorkersControlComponent,
-    PageNotFoundComponent
-
+    TasksComponent,
+    JobsComponent,
+    AppComponent,
+    NavComponent
   ],
   imports: [
-    BrowserModule,
-    AppRoutingModule,
+    MatProgressSpinnerModule,
     BrowserAnimationsModule,
-    MatButtonModule,
-    MatCheckboxModule,
-    LayoutModule,
-    MatToolbarModule,
-    MatSidenavModule,
-    MatIconModule,
-    MatListModule,
+    ReactiveFormsModule,
+    MatFormFieldModule,
+    MatPaginatorModule,
     MatGridListModule,
+    MatCheckboxModule,
+    AppRoutingModule,
+    MatSidenavModule,
+    MatToolbarModule,
+    HttpClientModule,
+    MatStepperModule,
+    MatButtonModule,
+    MatSelectModule,
+    MatDialogModule,
+    MatInputModule,
+    MatChipsModule,
+    MatTableModule,
+    MatSortModule,
+    BrowserModule,
+    MatListModule,
+    MatIconModule,
     MatCardModule,
     MatMenuModule,
-    MatTableModule,
-    MatPaginatorModule,
-    MatSortModule,
-    HttpClientModule,
+    LayoutModule,
     FormsModule,
-    MatFormFieldModule,
-    MatInputModule,
-    MatSelectModule,
-    MatStepperModule,
-    ReactiveFormsModule,
-    MatDialogModule,
-    MatChipsModule,
-    MatProgressSpinnerModule,
     CommonModule,
     ToastrModule.forRoot({
       timeOut: 3000,
@@ -94,6 +111,14 @@ import { CommonModule } from '@angular/common';
       easing: 'ease-out',
       easeTime: 300,
       progressBar: true
+    }),
+    TranslateModule.forRoot({
+      loader: {
+        provide: TranslateLoader,
+        useFactory: HttpLoaderFactory,
+        useClass: AccountTranslationLoader,
+        deps: [HttpClient],
+      },
     })
   ],
   exports: [
@@ -109,4 +134,9 @@ import { CommonModule } from '@angular/common';
   bootstrap: [AppComponent]
 })
 export class AppModule {
+
+}
+
+export function HttpLoaderFactory(http: HttpClient) {
+  return new TranslateHttpLoader(http);
 }
