@@ -1,25 +1,32 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
-import {ToastrModule} from 'ngx-toastr';
 import {AuthRoutingModule} from './auth-routing.module';
 import {LoginComponent} from './login/login.component';
 import {
   MatButtonModule,
-  MatCheckboxModule,
-  MatToolbarModule,
-  MatSidenavModule,
-  MatIconModule,
-  MatFormFieldModule,
-  MatListModule,
-  MatGridListModule,
   MatCardModule,
-  MatMenuModule, MatTableModule, MatPaginatorModule, MatSortModule, MAT_LABEL_GLOBAL_OPTIONS,
+  MatCheckboxModule,
+  MatChipsModule,
+  MatDialogModule,
+  MatFormFieldModule,
+  MatGridListModule,
+  MatIconModule,
   MatInputModule,
+  MatListModule,
+  MatMenuModule,
+  MatPaginatorModule,
   MatProgressSpinnerModule,
-  MatSelectModule, MatDialogModule, MatChipsModule, MatStepperModule
+  MatSelectModule,
+  MatSidenavModule,
+  MatSortModule,
+  MatStepperModule,
+  MatTableModule,
+  MatToolbarModule
 } from '@angular/material';
-import {HttpClientModule} from '@angular/common/http';
+import {HttpClient, HttpClientModule} from '@angular/common/http';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
+import {TranslateLoader, TranslateModule} from '@ngx-translate/core';
+import {AccountTranslationLoader, HttpLoaderFactory} from '../app.module';
 
 @NgModule({
   declarations: [LoginComponent],
@@ -41,7 +48,15 @@ import {FormsModule, ReactiveFormsModule} from '@angular/forms';
     MatProgressSpinnerModule,
     HttpClientModule,
     FormsModule,
-    ReactiveFormsModule
+    ReactiveFormsModule,
+    TranslateModule.forChild({
+      loader: {
+        provide: TranslateLoader,
+        useFactory: HttpLoaderFactory,
+        useClass: AccountTranslationLoader,
+        deps: [HttpClient],
+      },
+    })
   ],
 })
 export class AuthModule {
