@@ -1,5 +1,5 @@
 import {Component} from '@angular/core';
-
+import {TranslateService} from '@ngx-translate/core';
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
@@ -7,7 +7,11 @@ import {Component} from '@angular/core';
 })
 export class AppComponent {
 
-  constructor() {
+  constructor(private translate: TranslateService) {
+    translate.setDefaultLang('ru');
+  }
 
+  switchLanguage(language: string) {
+    this.translate.use(language);
   }
 }
